+++ src/pymmcore_widgets/_stack_viewer_v2/__init__.py
@@ -1,4 +1,3 @@
 from ._mda_viewer import MDAViewer
-from ._stack_viewer import StackViewer
 
-__all__ = ["StackViewer", "MDAViewer"]
+__all__ = ["MDAViewer"]
